requirements.txt
pyyaml
jsonschema
hashids
bleach
rfc3987
jinja2
pillow
aws-requests-auth
requests_aws4auth
elasticsearch
pyjwt
pycrypto
requests
requests_oauthlib
